Export all EGS related code to a static EgsTool class to be able to test it easily.
